# src/api/common/user/userController.js
/*
 * Copyright (c) Akveo 2019. All Rights Reserved.
 * Licensed under the Single Application / Multi Application License.
 * See LICENSE_SINGLE_APP / LICENSE_MULTI_APP in the 'docs' folder for license information on type of purchased license.
 */

const express = require('express');

const router = express.Router();
const { adminGuard } = require('../auth/aclService');
const UserService = require('./userService');
const CustomErrorService = require('../../../utils/customErrorService');
var path = require('path');
var root_path = path.dirname(require.main.filename);
var models = require(root_path + '/models');
var moment = require("moment");
var Moment = require('moment-timezone');

const userService = new UserService();

router.get('/', adminGuard, (req, res) => {
  userService
    .list(req.query)
    .then(users => res.send(users));
});

router.post('/', adminGuard, (req, res) => {
  userService
    .addUser(req.body)
    .then(user => res.send(user))
    .catch(err => res.status(400).send({ error: err.message }));
});

router.get('/current', (req, res) => {
  userService
    .findById(req.user.id)
    .then(user => {
      models.companies.findOne({
        where:{
          user_id : user.id
        }
      }).then(function(companies){
        user.company_id = companies ? companies.id : ''
        user.background_information = ((user.role == 'user') ? user.background_information : '');
        if(user.internship_preferences){
          user.internship_preferences = ((user.role == 'user') ? user.internship_preferences : '');
          user.internship_preferences.joinDate = ((user.role == 'user' && user.internship_preferences != null) ? Moment(new Date(user.internship_preferences.joinDate)).tz('Asia/Kolkata').format('YYYY-MM-DD') : '');
        }
        //console.log("user====>"+JSON.stringify(user))
        res.send(user)
      })
    });
});

router.put('/current', (req, res) => {
  userService
    .editCurrentUser(req.body, req.user.id)
    .then(user => res.send(user))
    .catch(error => {
      if (error instanceof CustomErrorService) {
        res.status(error.metadata && error.metadata.error.code)
          .send(error);
      }
    });
});

router.get('/:id', adminGuard, (req, res) => {
  userService
    .findById(req.params.id)
    .then(user => res.send(user));
});

router.put('/:id', adminGuard, (req, res) => {
  userService
    .editUser(req.body, req.params.id)
    .then(user => res.send(user))
    .catch(error => {
      if (error instanceof CustomErrorService) {
        res.status(error.metadata && error.metadata.error.code)
          .send(error);
      }
    });
});

router.delete('/:id', adminGuard, (req, res) => {
  userService
    .deleteUser(req.params.id)
    .then(() => res.send({ id: req.params.id }));
});

module.exports = router;